✨ rearborescence du projet + changement du fonctionnement de login et du context
