+++ package.json
@@ -10,11 +10,13 @@
     "@fullcalendar/list": "^5.6.0",
     "@fullcalendar/react": "^5.6.0",
     "@fullcalendar/timegrid": "^5.6.0",
+    "@material-ui/core": "^4.11.4",
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.8.1",
     "axios": "^0.21.1",
     "chart.js": "^2.9.4",
+    "mui-datatables": "^3.7.7",
     "npm": "^7.8.0",
     "react": "^17.0.1",
     "react-chartjs-2": "^2.11.1",
