--- src/components/Clients/DeleteClient.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function DeleteClient() {
-
-  return(
-    <div>
-      <h1>DeleteClient Page</h1>
-    </div>
-  );
-}
-export default DeleteClient;
