+++ src/components/Clients/ListClient.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function ListClient() {
+
+  return(
+    <div>
+      <h1>ClientList Page</h1>
+    </div>
+  );
+}
+export default ListClient;
