+++ src/components/navigation/sideMenu/SideMenu.css
@@ -20,7 +20,7 @@
     height: 130px;
     border-radius: 50%;
     background-color: rgb(243, 244, 246);
-    background-image: url('../../../assets/img/IcoLogo.png');
+    background-image: url('../../../assets/img/icoLogo.png');
     background-repeat: no-repeat;
     background-size: 55%;
     background-position: center;
