--- src/services/methode/auth-header.js
@@ -1,10 +1,0 @@
-export default function authHeader() {
-
-    const user = JSON.parse(localStorage.getItem('user'));
-
-    if (user && user.token) {
-        return { Authorization: 'Bearer ' + user.token };
-    } else {
-        return {};
-    }
-}
