+++ src/shared/services/appointmentType.service.js
@@ -1,0 +1,14 @@
+import instance from "./instanceAxios";
+
+
+class AppointmentTypeService {
+
+    path = '/appointmentTypes';
+    
+
+    getAll() {
+        return instance.get(`${this.path}`);
+    }
+}
+
+export default new AppointmentTypeService();
