+++ src/shared/services/clientType.service.js
@@ -1,0 +1,12 @@
+import instance from "./instanceAxios";
+
+class ClientTypeService {
+
+    path = '/clientTypes';
+
+    getAll() {
+        return instance.get(`${this.path}`);
+    }
+}
+
+export default new ClientTypeService();
