+++ src/shared/services/heaterType.service.js
@@ -1,0 +1,13 @@
+import instance from "./instanceAxios";
+
+
+class HeaterTypeService {
+
+    path = '/heaterTypes';
+
+    getAll() {
+        return instance.get(`${this.path}`);
+    }
+}
+
+export default new HeaterTypeService();
