+++ src/shared/services/propertyType.service.js
@@ -1,0 +1,13 @@
+import instance from "./instanceAxios";
+
+
+class PropertyTypeService {
+
+    path = '/propertyTypes';
+
+    getAll() {
+        return instance.get(`${this.path}`);
+    }
+}
+
+export default new PropertyTypeService();
