+++ src/shared/services/shutterType.service.js
@@ -1,0 +1,13 @@
+import instance from "./instanceAxios";
+
+
+class ShutterTypeService {
+
+    path = '/shutterTypes';
+
+    getAll() {
+        return instance.get(`${this.path}`);
+    }
+}
+
+export default new ShutterTypeService();
