+++ src/shared/services/userType.service.js
@@ -1,0 +1,12 @@
+import instance from "./instanceAxios";
+
+class UserTypeService {
+
+    path = '/userTypes';
+
+    getAll() {
+        return instance.get(`${this.path}`);
+    }
+}
+
+export default new UserTypeService();
